affichage des commentaires à la page d'accueil

+++ bcg-car-rentor-cd/backend/database/migrations/2025_02_22_200657_create_commentaires_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(User::class)->constrained()->cascadeOnDelete();
             $table->string("contenu");
-            $table->float("note");
+            $table->integer("note");
             $table->timestamps();
         });
     }
